refactor(header): address code review issues and fix circular import
- Extract queryClient singleton to src/lib/query-client.ts, breaking the
  AppShell -> header -> nav-config -> AppShell import cycle
- Update docker.tsx and nav-config.ts to import from the new module
- Tighten NoMenuNavItem.to to Exclude<RouteKey, MenuRouteKey> so the
  discriminated union enforces that no-menu items cannot use menu routes
- Replace &mdash; in DemoBanner with a plain hyphen (no em dash per style guide)
- Fix test imports outside __tests__/ to use @/ aliases instead of relative paths
- Update test mock targets from @/components/AppShell to @/lib/query-client
