+++ src/components/AppShell.tsx
@@ -1,5 +1,5 @@
 import { useEffect } from 'react'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { QueryClientProvider } from '@tanstack/react-query'
 import {
   DOCKER_PRELOAD_KEY, ZFS_PRELOAD_KEY, PROXMOX_PRELOAD_KEY,
   PRELOAD_STALE_TIME,
@@ -10,6 +10,7 @@ import Header from '@/components/header'
 import Toasts from '@/components/Toasts'
 import { useSettingsSync } from '@/hooks/useSettingsSync'
 import { useLightPaletteEffect } from '@/hooks/useLightPaletteEffect'
+import { queryClient } from '@/lib/query-client'
 
 if (import.meta.env.VITE_DEMO_MODE === 'true' && typeof window !== 'undefined') {
   // Use .then() instead of top-level await to avoid circular dependency deadlock.
@@ -19,8 +20,6 @@ if (import.meta.env.VITE_DEMO_MODE === 'true' && typeof window !== 'undefined') 
   // and MockEventSource._start uses setTimeout(50ms) - both fire after .then() resolves.
   import('@/lib/mock/install-demo').then(({ installDemo }) => installDemo()).catch((err) => console.error('Failed to install demo mode:', err))
 }
-
-export const queryClient = new QueryClient()
 
 export default function AppShell({ children }: { children: React.ReactNode }) {
   useSettingsSync()
